Move Sequential into tfe.keras.models
