+++ tf_encrypted/keras/__init__.py
@@ -2,12 +2,14 @@
 from __future__ import absolute_import
 
 from tf_encrypted.keras import engine
+from tf_encrypted.keras import models
 from tf_encrypted.keras import layers
-from tf_encrypted.keras.engine.sequential import Sequential
+from tf_encrypted.keras.models import Sequential
 
 
 __all__ = [
     'engine',
+    'models',
     'layers',
     'Sequential',
 ]
