+++ tf_encrypted/keras/models/__init__.py
@@ -1,0 +1,9 @@
+"""Keras models in TF Encrypted."""
+from __future__ import absolute_import
+
+from .sequential import Sequential
+
+
+__all__ = [
+    'Sequential',
+]
